+++ app.js
@@ -2,12 +2,11 @@ const express = require('express');
 const cors = require('cors');
 
 const userRouter = require('./routes/users/users');
-const eventRouter = require('./routes/events');
+const eventRouter = require('./routes/events/events');
 const hoursRouter = require('./routes/volunteer_hours');
 const posteventRouter = require('./routes/postevents');
 const volunteerRouter = require('./routes/volunteers');
 const availabilityRouter = require('./routes/users/availability');
-const driverRouter = require('./routes/drivers');
 
 const app = express();
 const PORT = process.env.PORT || 3001;
@@ -28,7 +27,6 @@ app.use('/postevents', posteventRouter);
 app.use('/events', eventRouter);
 app.use('/users', userRouter);
 app.use('/availability', availabilityRouter);
-app.use('/drivers', driverRouter);
 
 app.listen(PORT, () => {
   console.log(`Server listening on ${PORT}`);
