+++ routes/events/events.js
@@ -1,7 +1,7 @@
 // endpoints related to events
 const express = require('express');
-const { pool, db } = require('../server/db');
-const { isNumeric, isZipCode, keysToCamel } = require('./utils');
+const { pool, db } = require('../../server/db');
+const { isNumeric, isZipCode, keysToCamel } = require('../utils');
 
 const eventRouter = express();
 
@@ -57,28 +57,31 @@ eventRouter.get('/', async (req, res) => {
 });
 
 // get all upcoming events
-eventRouter.get('/upcoming', async (req, res) => {
-  try {
-    const currDate = new Date();
-    const conditions = 'WHERE start_datetime >= $1';
-    const events = await pool.query(getEventsQuery(conditions), [currDate]);
-    res.status(200).json(keysToCamel(events.rows));
-  } catch (err) {
-    res.status(400).send(err.message);
-  }
-});
+// dont think we need this because event filtering should be done in the frontend
+// just need to use the get all events endpoint
+// eventRouter.get('/upcoming', async (req, res) => {
+//   try {
+//     const currDate = new Date();
+//     const conditions = 'WHERE start_datetime >= $1';
+//     const events = await pool.query(getEventsQuery(conditions), [currDate]);
+//     res.status(200).json(keysToCamel(events.rows));
+//   } catch (err) {
+//     res.status(400).send(err.message);
+//   }
+// });
 
 // get all past events
-eventRouter.get('/past', async (req, res) => {
-  try {
-    const currDate = new Date();
-    const conditions = 'WHERE start_datetime < $1';
-    const events = await pool.query(getEventsQuery(conditions), [currDate]);
-    res.status(200).json(keysToCamel(events.rows));
-  } catch (err) {
-    res.status(400).send(err.message);
-  }
-});
+// dont think we need this because event filtering should be done in the frontend
+// eventRouter.get('/past', async (req, res) => {
+//   try {
+//     const currDate = new Date();
+//     const conditions = 'WHERE start_datetime < $1';
+//     const events = await pool.query(getEventsQuery(conditions), [currDate]);
+//     res.status(200).json(keysToCamel(events.rows));
+//   } catch (err) {
+//     res.status(400).send(err.message);
+//   }
+// });
 
 // get an event
 eventRouter.get('/:eventId', async (req, res) => {
